+++ src/columns/uid_login.rs
@@ -3,6 +3,8 @@ use crate::{column_default, Column};
 use std::cmp;
 use std::collections::HashMap;
 
+const UID_NOT_SET: u32 = 0xffffffff;
+
 pub struct UidLogin {
     header: String,
     unit: String,
@@ -28,7 +30,11 @@ impl UidLogin {
 impl Column for UidLogin {
     fn add(&mut self, proc: &ProcessInfo) {
         let (fmt_content, raw_content) = if let Ok(uid) = proc.curr_proc.loginuid() {
-            (format!("{uid}"), uid)
+            if uid == UID_NOT_SET {
+                (String::new(), uid)
+            } else {
+                (format!("{uid}"), uid)
+            }
         } else {
             (String::new(), 0)
         };